📝 some explanation
